.agent/requirements.txt
python-dotenv
pydantic
openai
google-generativeai
pytest
